move methods only need by extended
